refactor: Reorganize JAX handlers module structure
