+++ effectful/handlers/jax/__init__.py
@@ -6,10 +6,10 @@ except ImportError:
     raise ImportError("Jax and Numpyro are required to use effectful.handlers.jax")
 
 # side effect: register defdata for jax.Array
-import effectful.handlers.jax.terms  # noqa: F401
+import effectful.handlers.jax._terms  # noqa: F401
 
-from .handlers import bind_dims as bind_dims
-from .handlers import jax_getitem as jax_getitem
-from .handlers import jit as jit
-from .handlers import sizesof as sizesof
-from .handlers import unbind_dims as unbind_dims
+from ._handlers import bind_dims as bind_dims
+from ._handlers import jax_getitem as jax_getitem
+from ._handlers import jit as jit
+from ._handlers import sizesof as sizesof
+from ._handlers import unbind_dims as unbind_dims
